scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
azure-core==1.29.4
    # via azure-identity
azure-identity==1.14.0
    # via -r requirements.in
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
cryptography==41.0.4
    # via
    #   azure-identity
    #   msal
    #   pyjwt
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
msal==1.24.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.28.1
    # via -r requirements.in
portalocker==2.8.2
    # via msal-extensions
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via msal
requests==2.31.0
    # via
    #   azure-core
    #   msal
    #   openai
six==1.16.0
    # via azure-core
tenacity==8.2.3
    # via -r requirements.in
tqdm==4.66.1
    # via openai
typing-extensions==4.8.0
    # via azure-core
urllib3==2.0.7
    # via requests
yarl==1.9.2
    # via aiohttp
